+++ MsfsPlugin/MsfsPlugin/msfs/MsfsData.cs
@@ -31,7 +31,7 @@
         private Int16 currentRudderTrim;       
         private Int32 currentGearHandle;
 
-        private Int16 currentZoom;
+        private Int32 currentZoom;
 
         private Boolean connected; 
         private Boolean tryingToconnect;
@@ -94,8 +94,8 @@
         public Int16 CurrentAileronTrimFromMSFS { get => this.currentAileronTrim; set { this.currentAileronTrim = value; } }
         public Int16 CurrentRudderTrim { get => this.currentRudderTrim; set { this.currentRudderTrim = value; this.SetToMSFS = true; } }
         public Int16 CurrentRudderTrimFromMSFS { get => this.currentRudderTrim; set { this.currentRudderTrim = value; } }
-        public Int16 CurrentZoom { get => this.currentZoom; set { this.currentZoom = value; this.SetToMSFS = true; } }
-        public Int16 CurrentZoomFromMSFS { get => this.currentZoom; set { this.currentZoom = value; } }
+        public Int32 CurrentZoom { get => this.currentZoom; set { this.currentZoom = value; this.SetToMSFS = true; } }
+        public Int32 CurrentZoomFromMSFS { get => this.currentZoom; set { this.currentZoom = value; } }
         public Boolean Connected { get => this.connected; set { this.connected = value; this.changed(); } }
         public Boolean TryConnect { get => this.tryingToconnect; set { this.tryingToconnect = value; this.changed(); } }
         public Boolean ValuesDisplayed { get => this.valuesDisplayed; set { this.valuesDisplayed = value; SimulatorDAO.Initialise(); } }
